requirements.txt
certifi==2021.10.8
click==8.1.3
colorama==0.4.4
cycler==0.11.0
Flask==2.1.2
Flask-Cors==3.0.10
fonttools==4.33.3
geomet==0.2.1.post1
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.1.0
kiwisolver==1.4.2
kneed==0.7.0
MarkupSafe==2.1.1
matplotlib==3.5.2
mysql-connector-python==8.0.29
numpy==1.21.6
packaging==21.3
pandas==1.3.5
Pillow==9.1.0
protobuf==3.20.1
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
sklearn==0.0
threadpoolctl==3.1.0
typing_extensions==4.2.0
Werkzeug==2.1.2
wincertstore==0.2
xgboost==0.90
zipp==3.8.0
